requirements.txt
Django==4.2
Pillow==9.5.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
djangorestframework==3.14.0
drf-yasg==1.21.6
django-filter==23.2
coreapi==2.3.3
